Separated methods to track visits and list visits
