+++ module/Rest/config/dependencies.config.php
@@ -66,8 +66,8 @@ return [
             Service\ShortUrl\ShortUrlResolver::class,
             ShortUrlDataTransformer::class,
         ],
-        Action\Visit\ShortUrlVisitsAction::class => [Service\VisitsTracker::class],
-        Action\Visit\TagVisitsAction::class => [Service\VisitsTracker::class],
+        Action\Visit\ShortUrlVisitsAction::class => [Visit\VisitsStatsHelper::class],
+        Action\Visit\TagVisitsAction::class => [Visit\VisitsStatsHelper::class],
         Action\Visit\GlobalVisitsAction::class => [Visit\VisitsStatsHelper::class],
         Action\ShortUrl\ListShortUrlsAction::class => [Service\ShortUrlService::class, ShortUrlDataTransformer::class],
         Action\ShortUrl\EditShortUrlTagsAction::class => [Service\ShortUrlService::class],
