+++ module/Rest/src/Action/Visit/ShortUrlVisitsAction.php
@@ -10,7 +10,7 @@ use Psr\Http\Message\ServerRequestInterface as Request;
 use Shlinkio\Shlink\Common\Paginator\Util\PagerfantaUtilsTrait;
 use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\Model\VisitsParams;
-use Shlinkio\Shlink\Core\Service\VisitsTrackerInterface;
+use Shlinkio\Shlink\Core\Visit\VisitsStatsHelperInterface;
 use Shlinkio\Shlink\Rest\Action\AbstractRestAction;
 use Shlinkio\Shlink\Rest\Middleware\AuthenticationMiddleware;
 
@@ -21,11 +21,11 @@ class ShortUrlVisitsAction extends AbstractRestAction
     protected const ROUTE_PATH = '/short-urls/{shortCode}/visits';
     protected const ROUTE_ALLOWED_METHODS = [self::METHOD_GET];
 
-    private VisitsTrackerInterface $visitsTracker;
+    private VisitsStatsHelperInterface $visitsHelper;
 
-    public function __construct(VisitsTrackerInterface $visitsTracker)
+    public function __construct(VisitsStatsHelperInterface $visitsHelper)
     {
-        $this->visitsTracker = $visitsTracker;
+        $this->visitsHelper = $visitsHelper;
     }
 
     public function handle(Request $request): Response
@@ -33,7 +33,7 @@ class ShortUrlVisitsAction extends AbstractRestAction
         $identifier = ShortUrlIdentifier::fromApiRequest($request);
         $params = VisitsParams::fromRawData($request->getQueryParams());
         $apiKey = AuthenticationMiddleware::apiKeyFromRequest($request);
-        $visits = $this->visitsTracker->info($identifier, $params, $apiKey);
+        $visits = $this->visitsHelper->visitsForShortUrl($identifier, $params, $apiKey);
 
         return new JsonResponse([
             'visits' => $this->serializePaginator($visits),
