+++ module/Rest/src/Action/Visit/TagVisitsAction.php
@@ -9,7 +9,7 @@ use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Shlinkio\Shlink\Common\Paginator\Util\PagerfantaUtilsTrait;
 use Shlinkio\Shlink\Core\Model\VisitsParams;
-use Shlinkio\Shlink\Core\Service\VisitsTrackerInterface;
+use Shlinkio\Shlink\Core\Visit\VisitsStatsHelperInterface;
 use Shlinkio\Shlink\Rest\Action\AbstractRestAction;
 use Shlinkio\Shlink\Rest\Middleware\AuthenticationMiddleware;
 
@@ -20,11 +20,11 @@ class TagVisitsAction extends AbstractRestAction
     protected const ROUTE_PATH = '/tags/{tag}/visits';
     protected const ROUTE_ALLOWED_METHODS = [self::METHOD_GET];
 
-    private VisitsTrackerInterface $visitsTracker;
+    private VisitsStatsHelperInterface $visitsHelper;
 
-    public function __construct(VisitsTrackerInterface $visitsTracker)
+    public function __construct(VisitsStatsHelperInterface $visitsHelper)
     {
-        $this->visitsTracker = $visitsTracker;
+        $this->visitsHelper = $visitsHelper;
     }
 
     public function handle(Request $request): Response
@@ -32,7 +32,7 @@ class TagVisitsAction extends AbstractRestAction
         $tag = $request->getAttribute('tag', '');
         $params = VisitsParams::fromRawData($request->getQueryParams());
         $apiKey = AuthenticationMiddleware::apiKeyFromRequest($request);
-        $visits = $this->visitsTracker->visitsForTag($tag, $params, $apiKey);
+        $visits = $this->visitsHelper->visitsForTag($tag, $params, $apiKey);
 
         return new JsonResponse([
             'visits' => $this->serializePaginator($visits),
